Extract and rename a bunch of classes
