+++ Xamarin.PropertyEditing.Mac/Controls/Custom/ChannelEditorType.cs
@@ -1,0 +1,10 @@
+﻿namespace Xamarin.PropertyEditing.Mac
+{
+	public enum ChannelEditorType
+	{
+		RGB,
+		HLS,
+		HSB,
+		CMYK
+	};
+}
