+++ Xamarin.PropertyEditing.Mac/Controls/Custom/MaterialColorLayer.cs
@@ -1,0 +1,66 @@
+﻿using CoreAnimation;
+using Xamarin.PropertyEditing.Drawing;
+
+namespace Xamarin.PropertyEditing.Mac
+{
+	class MaterialColorLayer : CATextLayer
+	{
+		public MaterialColorLayer ()
+		{
+			AddSublayer (Selection);
+		}
+
+		CATextLayer Selection { get; } = new CATextLayer () {
+			CornerRadius = 3
+		};
+
+		string text;
+		public string Text
+		{
+			get => text;
+			set
+			{
+				text = value;
+				SetNeedsLayout ();
+			}
+		}
+
+		CommonColor materialColor;
+		public new CommonColor BackgroundColor
+		{
+			get => materialColor;
+			set
+			{
+				materialColor = value;
+				base.BackgroundColor = materialColor.ToCGColor ();
+			}
+		}
+
+		bool isSelected;
+		public bool IsSelected
+		{
+			get => isSelected;
+			set
+			{
+				if (isSelected == value)
+					return;
+				isSelected = value;
+				SetNeedsLayout ();
+			}
+		}
+
+		public override void LayoutSublayers ()
+		{
+			base.LayoutSublayers ();
+			//String = isSelected ? "" : text;
+			Selection.String = text;
+			Selection.Frame = Frame.Bounds ().Border (new CommonThickness (3));
+			Selection.BorderWidth = isSelected ? 2 : 0;
+			Selection.BorderColor = ForegroundColor;
+			Selection.ForegroundColor = ForegroundColor;
+			Selection.FontSize = FontSize;
+			Selection.ContentsScale = ContentsScale;
+			Selection.TextAlignmentMode = TextAlignmentMode;
+		}
+	}
+}
